Split interface
